requirements.txt
numpy>=1.21.0
pandas>=1.3.0
ccxt>=2.0.0
python-dotenv>=0.19.0
websocket-client>=1.2.1
requests>=2.26.0
ta>=0.7.0
matplotlib>=3.4.3
scipy>=1.7.0
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.0
eth_account>=0.5.9
aiohttp>=3.8.1
asyncio>=3.4.3
colorama>=0.4.4
schedule>=1.1.0
websockets>=10.0
